this isn't related to this branch but rename FlagMaker class file to be PascalCase
