+++ src/FlagMaker.js
